Quiz: Count Unique Words
Your task for this quiz is to find the number of unique words in the text. In the code editor below, complete these three steps to get your answer.

Split verse into a list of words. Hint: You can use a string method you learned in the previous lesson.
Convert the list into a data structure that would keep only the unique elements from the list.
Print the length of the container.