Quiz: Verse Dictionary
In the code editor below, you'll find a dictionary containing the unique words of verse stored as keys and the number of times they appear in verse stored as values. Use this dictionary to answer the following questions. Submit these answers in the quiz below the code editor.

How many unique words are in verse_dict? Is the key "breathe" in verse_dict?
What is the first element in the list created when verse_dict is sorted by keys?
Which key (word) has the highest value in verse_dict?